refactor microservice requests to helper functions in utils
