+++ refactor/dispatcher/app/api/api_v0/endpoints/workflow.py
@@ -33,6 +33,7 @@ from app.core.dispatch_workflow import (
 )
 from app.core.dispatcher_logger import logger
 from app.core.update_workflow import update_workflow_results
+from app.core.utils import update_result_and_ui
 from app.schemas.workflow import (
     CancelWorkflowResponse,
     DispatchWorkflowResponse,
@@ -89,11 +90,6 @@ mock_result = {
         },
     },
 }
-
-
-def get_result(dispatch_id: str):
-    resp = requests.get(f"{BASE_URI}/api/v0/workflow/results/{dispatch_id}")
-    return resp.content
 
 
 logger.warning("Dispatcher Service Started")
@@ -163,8 +159,6 @@ def update_workflow(
         workflow_status_queue.get()  # Empty queue when workflow is no longer running (completed
         # or failed)
 
-    send_result_object_to_result_service(updated_result_obj)
-
-    send_task_update_to_ui(dispatch_id=dispatch_id, task_id=task_id)
+    update_result_and_ui()
 
     return {"response": f"{dispatch_id} workflow updated successfully"}
